AggregationField dragging data transfer [wip]
