+++ src/scss/pipelines.scss
@@ -124,7 +124,7 @@ $pipeline-width: 180px;
     .transform-item-enum {
       margin: 0;
     }
-    
+
     .transform-item,
     .transform-item-enum:not(:last-child) {
       border-bottom: 1px solid #000;
